+++ lib/app/app.dart
@@ -1,15 +1,14 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chatbot/app/models/theme_model.dart';
 import 'package:flutter_chatbot/app/services/firebase_auth_service.dart';
-import 'package:flutter_chatbot/ui/views/authentication/create_account_view.dart';
+import 'package:flutter_chatbot/ui/views/auth/create_account_view.dart';
 import 'package:flutter_chatbot/ui/views/onboarding/onboard_pageview.dart';
 import 'package:flutter_chatbot/app/routing/private_route.dart';
 import 'package:flutter_chatbot/app/routing/public_route.dart';
 import 'package:provider/provider.dart';
 import '../app/constants/strings.dart';
-import '../ui/views/authentication/sign_in_view.dart';
-import '../ui/views/messaging/messaging_view.dart';
+import 'package:flutter_chatbot/ui/views/auth/sign_in_view.dart';
+import 'package:flutter_chatbot/ui/views/messaging/messaging_view.dart';
 
 class MyApp extends StatelessWidget {
   @override
@@ -38,37 +37,3 @@ class MyApp extends StatelessWidget {
         }));
   }
 }
-
-class LandingPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<User>(
-      stream: FirebaseAuth.instance.authStateChanges(),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.active) {
-          User user = snapshot.data;
-          if (user == null) {
-            return SignInView();
-          }
-          return MessagingView();
-        } else {
-          return Scaffold(
-            body: Center(
-              child: CircularProgressIndicator(),
-            ),
-          );
-        }
-      },
-    );
-  }
-}
-
-// return MaterialApp(
-//       title: 'Material App',
-//       theme: themeNotifier.getTheme(),
-//       initialRoute: '/',
-//       routes: {
-//         Strings.homeRoute: (context) => SignInView(),
-//         Strings.messsagingViewRoute: (context) => MessagingView(),
-//       },
-//     );
