+++ lib/app/constants/onboarding.dart
@@ -1,0 +1,22 @@
+class Onboarding {
+  static const List<String> assets = [
+    "assets/images/onboardTalk.png",
+    "assets/images/onboardFeedback.png",
+    "assets/images/onboardModes.png",
+    "assets/images/onboardTalk.png"
+  ];
+
+  static const List<String> headers = [
+    "Talk with ChatBot",
+    "Provide feedback",
+    "Switch Modes",
+    "ChatBot is designed for you"
+  ];
+
+  static const List<String> body = [
+    "Chat about how you've been feeling and what has been on your mind",
+    "Let ChatBot know what it does well and/or how it can do better",
+    "Choose between dark and light mode",
+    "ChatBot is here to listen to all your thoughts and feelings, so don't be afraid to share"
+  ];
+}
