+++ lib/ui/views/auth/sign_in_view.dart
@@ -1,56 +1,39 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_chatbot/app/constants/strings.dart';
+import 'package:flutter_chatbot/ui/page_base.dart';
 import 'package:flutter_chatbot/ui/widgets/auth/core/auth_view_header.dart';
 import 'package:flutter_chatbot/ui/widgets/auth/forms/sign_in_form.dart';
+import 'package:flutter_chatbot/ui/widgets/auth/buttons/google_sign_in_button.dart';
+import 'package:flutter_chatbot/ui/widgets/auth/buttons/go_to_button.dart';
+import 'package:flutter_chatbot/app/constants/strings.dart';
 import 'package:flutter_svg/svg.dart';
-import '../../widgets/auth/core/google_sign_in_button.dart';
 
 class SignInView extends StatelessWidget {
   const SignInView({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      color: Theme.of(context).backgroundColor,
-      child: Scaffold(
-        backgroundColor: Theme.of(context).backgroundColor,
-        body: SingleChildScrollView(
-          child: ConstrainedBox(
-              constraints: BoxConstraints(
-                  maxWidth: MediaQuery.of(context).size.width,
-                  maxHeight: MediaQuery.of(context).size.height),
-              child: SafeArea(
-                  child: Center(
-                      child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  AuthViewHeader(),
-                  SignInForm(),
-                  Container(
-                    margin: EdgeInsets.only(top: 20),
-                    child: SvgPicture.asset("assets/images/OR.svg"),
-                  ),
-                  GoogleSignInButton(),
-                  Spacer(),
-                  Container(
-                      margin: EdgeInsets.only(bottom: 20),
-                      child: FlatButton(
-                        onPressed: () {
-                          Navigator.pushNamed(
-                              context, Strings.createAccountRoute);
-                        },
-                        child: Text(
-                          "New user? Create an account",
-                          style: Theme.of(context)
-                              .textTheme
-                              .bodyText1
-                              .copyWith(color: Colors.grey),
-                        ),
-                      )),
-                ],
-              )))),
+    void navigate() {
+      Navigator.pushNamed(context, Strings.createAccountRoute);
+    }
+
+    return PageBase(
+        childWidget: Center(
+            child: Column(
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        AuthViewHeader(),
+        SignInForm(),
+        Container(
+          margin: EdgeInsets.only(top: 20),
+          child: SvgPicture.asset("assets/images/OR.svg"),
         ),
-      ),
-    );
+        GoogleSignInButton(),
+        Spacer(),
+        GoToButton(
+          message: "New user? Create an account.",
+          navigate: navigate,
+        )
+      ],
+    )));
   }
 }
