+++ lib/ui/widgets/messaging/chatbot/bot_response.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chatbot/app/models/chat_model.dart';
 import 'package:flutter_chatbot/assets/assets.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/chat_nip.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/message_bubble.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/message/chat_nip.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/message/message_bubble.dart';
 import 'package:provider/provider.dart';
 
 class BotResponse extends StatelessWidget {
