+++ lib/ui/widgets/messaging/core/interactive_chat_window.dart
@@ -5,10 +5,10 @@ import 'package:flutter_chatbot/app/constants/strings.dart';
 import 'package:flutter_chatbot/app/models/chat_model.dart';
 import 'package:flutter_chatbot/app/services/firebase_db_service.dart';
 import 'package:flutter_chatbot/app/services/firebase_auth_service.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/avatar_view.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/chatbot/avatar_view.dart';
 import 'package:flutter_chatbot/ui/widgets/messaging/feedback/feedback_overlay.dart';
 import 'package:flutter_chatbot/ui/widgets/messaging/settings/settings_overlay.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/text_composer.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/core/text_composer.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
