+++ lib/ui/widgets/messaging/message/chat_message.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/chat_nip.dart';
-import 'package:flutter_chatbot/ui/views/messaging/widgets/message_bubble.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/message/chat_nip.dart';
+import 'package:flutter_chatbot/ui/widgets/messaging/message/message_bubble.dart';
 import 'package:flutter_chatbot/assets/assets.dart';
 
 class ChatMessage extends StatelessWidget {
